delete files in pj
